+++ lib/src/core/better_player_controller.dart
@@ -37,8 +37,6 @@ class BetterPlayerController extends ChangeNotifier {
   Widget Function(BuildContext context, String errorMessage) get errorBuilder =>
       betterPlayerConfiguration.errorBuilder;
 
-  double get aspectRatio => betterPlayerConfiguration.aspectRatio;
-
   Widget get placeholder => betterPlayerConfiguration.placeholder;
 
   Widget get overlay => betterPlayerConfiguration.overlay;
@@ -127,6 +125,8 @@ class BetterPlayerController extends ChangeNotifier {
   bool _controlsEnabled = true;
 
   bool get controlsEnabled => _controlsEnabled;
+
+  double _overriddenAspectRatio;
 
   BetterPlayerController(
     this.betterPlayerConfiguration, {
@@ -580,6 +580,20 @@ class BetterPlayerController extends ChangeNotifier {
         play();
       }
     }
+  }
+
+  ///Setup overridden aspect ratio.
+  void setOverriddenAspectRatio(double aspectRatio) {
+    _overriddenAspectRatio = aspectRatio;
+  }
+
+  ///Get aspect ratio used in current video. If aspect ratio is null, then
+  ///aspect ratio from BetterPlayerConfiguration will be used. Otherwise
+  ///[_overriddenAspectRatio] will be used.
+  double getAspectRatio() {
+    return _overriddenAspectRatio != null
+        ? _overriddenAspectRatio
+        : betterPlayerConfiguration.aspectRatio;
   }
 
   @override