+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.37
 * Added setControlsEnabled to BetterPlayerController
 * Fixed example video list widget buttons not rendering correctly in small resolutions
+* Added setOverriddenAspectRatio to BetterPlayerController.
 
 ## 0.0.36
 * Added INITIALIZED event